requirements.txt
# Core dependencies only
Appium-Python-Client>=3.0.0
pytest>=7.4.0
selenium>=4.15.0
